backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
bcrypt==4.2.0
pydantic[email]==2.9.2
python-dotenv==1.0.1
razorpay==2.0.1
python-multipart==0.0.9
